refactor the JSON extensions/wrapper to separate pakage within repo. Existing classes to be an extension of the new classes
