+++ .gitignore
@@ -3,3 +3,4 @@
 .settings/
 .project
 .classpath
+.idea/
