+++ src/main/java/in/succinct/beckn/BecknObjectWithId.java
@@ -1,9 +1,9 @@
 package in.succinct.beckn;
 
-import com.venky.core.util.ObjectUtil;
+import in.succinct.json.ExtendedJSONObjectWithId;
 import org.json.simple.JSONObject;
 
-public class BecknObjectWithId extends BecknObject {
+public class BecknObjectWithId extends ExtendedJSONObjectWithId {
     public BecknObjectWithId(){
         super();
     }
@@ -13,16 +13,5 @@ public class BecknObjectWithId extends BecknObject {
     public BecknObjectWithId(JSONObject object){
         super(object);
     }
-
-    public String getId(){
-        return get("id");
-    }
-    public void setId(String id){
-        if (!ObjectUtil.isVoid(id) && !ObjectUtil.isVoid(getId()) && !ObjectUtil.equals(getId(),id)){
-            throw new RuntimeException("Id cannot be changed!");
-        }
-        set("id",id);
-    }
-
 
 }
