+++ src/main/java/in/succinct/beckn/BecknStrings.java
@@ -1,10 +1,11 @@
 package in.succinct.beckn;
 
+import in.succinct.json.ExtendedJSONStrings;
 import org.json.simple.JSONArray;
 
-public class BecknStrings extends BecknObjects<String>{
-
+public class BecknStrings extends ExtendedJSONStrings {
     public BecknStrings() {
+        super();
     }
 
     public BecknStrings(JSONArray value) {
