+++ src/main/java/in/succinct/json/ExtendedJSONStrings.java
@@ -1,0 +1,13 @@
+package in.succinct.json;
+
+import org.json.simple.JSONArray;
+
+public class ExtendedJSONStrings extends ExtendedJSONObjects<String>{
+
+    public ExtendedJSONStrings() {
+    }
+
+    public ExtendedJSONStrings(JSONArray value) {
+        super(value);
+    }
+}
